chore: #66 - AlcoholLevelCard 내부로 이동
